Time: 0 ms (100.00%) | Memory: 7.2 MB (13.61%) - LeetSync

+++ 278-first-bad-version/first-bad-version.cpp
@@ -7,7 +7,7 @@ public:
         int l = 0,r = n,ans = 0;
         while(l <= r)
         {
-            long long int mid = l+(r-l)/2;
+            int mid = l+(r-l)/2;
             if(isBadVersion(mid))
             {
                 ans = mid;
